Add more functionality to market place page
The batch component now display more information and the buttons
in the component work. However, for now the Claim button is not connected to anything
but it is able to display if the batch is claimed or not.
By clicking the details button in the BatchDetails component, it shows up on the right side
and by clicking the button a second time it hides.